backend/requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pandas
numpy
icalendar
matplotlib
faiss-cpu
chroma
ollama
langchain
pdfplumber
tiktoken

